requirements.txt
numpy~=1.19.2
matplotlib~=3.3.2
scipy~=1.5.2
pandas~=1.1.3
yfinance~=0.1.87
requests~=2.28.1
beautifulsoup4~=4.9.3
python-dateutil~=2.8.1